refactor(user-user-store): 폴더 경로 변경한다 (#282)
